requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.84
botocore==1.27.84
Django==4.1.1
django-storages==1.13.1
djangorestframework==3.14.0
jmespath==1.0.1
pkg-resources==0.0.0
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
